+++ src/lib.rs
@@ -3,7 +3,10 @@ extern crate core;
 #[cfg(feature = "serde")]
 mod serde;
 
+mod iter;
+
 use core::iter::{IntoIterator, Iterator};
+use iter::*;
 
 #[derive(Clone)]
 pub struct IntMap<V> {
@@ -530,249 +533,6 @@ fn ensure_load_rate<V>(
     has_cache_increased
 }
 
-use std::slice::Iter as SliceIter;
-use std::slice::IterMut as SliceIterMut;
-use std::vec::IntoIter as VecIntoIter;
-
-// ***************** Iter *********************
-
-pub struct Iter<'a, K: 'a, V: 'a> {
-    outer: SliceIter<'a, Vec<(K, V)>>,
-    inner: SliceIter<'a, (K, V)>,
-}
-
-impl<'a, K, V> Iter<'a, K, V> {
-    pub fn new(vec: &'a Vec<Vec<(K, V)>>) -> Self {
-        let mut outer = vec.iter();
-        let inner = { outer.next() }
-            .map(|v| v.iter())
-            .unwrap_or_else(|| (&[]).iter());
-
-        Iter {
-            outer: outer,
-            inner: inner,
-        }
-    }
-}
-
-impl<'a, K, V> Iterator for Iter<'a, K, V> {
-    type Item = (&'a K, &'a V);
-
-    #[inline]
-    fn next(&mut self) -> Option<(&'a K, &'a V)> {
-        loop {
-            if let Some(r) = self.inner.next() {
-                return Some((&r.0, &r.1));
-            }
-
-            self.inner = self.outer.next()?.iter();
-        }
-    }
-}
-
-// ***************** Iter Mut *********************
-
-pub struct IterMut<'a, K: 'a, V: 'a> {
-    outer: SliceIterMut<'a, Vec<(K, V)>>,
-    inner: SliceIterMut<'a, (K, V)>,
-}
-
-impl<'a, K, V> IterMut<'a, K, V> {
-    fn new(vec: &'a mut Vec<Vec<(K, V)>>) -> IterMut<'a, K, V> {
-        let mut outer = vec.iter_mut();
-        let inner = { outer.next() }
-            .map(|v| v.iter_mut())
-            .unwrap_or_else(|| (&mut []).iter_mut());
-
-        IterMut {
-            outer: outer,
-            inner: inner,
-        }
-    }
-}
-
-impl<'a, K, V> Iterator for IterMut<'a, K, V> {
-    type Item = (&'a K, &'a mut V);
-
-    #[inline]
-    fn next(&mut self) -> Option<(&'a K, &'a mut V)> {
-        loop {
-            if let Some(r) = self.inner.next() {
-                return Some((&r.0, &mut r.1));
-            }
-
-            self.inner = self.outer.next()?.iter_mut();
-        }
-    }
-}
-
-// ***************** Values Iter *********************
-
-pub struct Values<'a, K: 'a, V: 'a> {
-    inner: Iter<'a, K, V>,
-}
-
-impl<'a, K, V> Iterator for Values<'a, K, V> {
-    type Item = &'a V;
-
-    #[inline]
-    fn next(&mut self) -> Option<&'a V> {
-        self.inner.next().map(|kv| kv.1)
-    }
-    #[inline]
-    fn size_hint(&self) -> (usize, Option<usize>) {
-        self.inner.size_hint()
-    }
-}
-
-// ***************** Keys Iter *********************
-
-pub struct Keys<'a, K: 'a, V: 'a> {
-    inner: Iter<'a, K, V>,
-}
-
-impl<'a, K, V> Iterator for Keys<'a, K, V> {
-    type Item = &'a K;
-
-    #[inline]
-    fn next(&mut self) -> Option<&'a K> {
-        self.inner.next().map(|kv| kv.0)
-    }
-    #[inline]
-    fn size_hint(&self) -> (usize, Option<usize>) {
-        self.inner.size_hint()
-    }
-}
-
-// ***************** Values Mut *********************
-
-pub struct ValuesMut<'a, K: 'a, V: 'a> {
-    inner: IterMut<'a, K, V>,
-}
-
-impl<'a, K, V> Iterator for ValuesMut<'a, K, V> {
-    type Item = &'a mut V;
-
-    #[inline]
-    fn next(&mut self) -> Option<&'a mut V> {
-        self.inner.next().map(|kv| kv.1)
-    }
-
-    #[inline]
-    fn size_hint(&self) -> (usize, Option<usize>) {
-        self.inner.size_hint()
-    }
-}
-
-pub struct Drain<'a, K: 'a, V: 'a> {
-    count: &'a mut usize,
-    outer: SliceIterMut<'a, Vec<(K, V)>>,
-    inner: Option<std::vec::Drain<'a, (K, V)>>,
-}
-
-impl<'a, K, V> Drain<'a, K, V> {
-    fn new(vec: &'a mut Vec<Vec<(K, V)>>, count: &'a mut usize) -> Drain<'a, K, V> {
-        let mut outer = vec.iter_mut();
-        let inner = outer
-            .next()
-            .map(|v| Some(v.drain(..)))
-            .unwrap_or_else(|| None);
-
-        Drain {
-            count: count,
-            outer: outer,
-            inner: inner,
-        }
-    }
-}
-
-impl<'a, K, V> Iterator for Drain<'a, K, V> {
-    type Item = (K, V);
-
-    #[inline]
-    fn next(&mut self) -> Option<(K, V)> {
-        loop {
-            if let Some(r) = self.inner.as_mut().and_then(|i| i.next()) {
-                *self.count -= 1;
-                return Some((r.0, r.1));
-            }
-
-            self.inner = Some(self.outer.next()?.drain(..));
-        }
-    }
-}
-
-// ***************** Into Iter *********************
-
-impl<V> IntoIterator for IntMap<V> {
-    type Item = (u64, V);
-    type IntoIter = IntoIter<u64, V>;
-
-    fn into_iter(self) -> Self::IntoIter {
-        IntoIter::new(self.cache)
-    }
-}
-
-pub struct IntoIter<K, V> {
-    outer: VecIntoIter<Vec<(K, V)>>,
-    inner: VecIntoIter<(K, V)>,
-}
-
-impl<K, V> IntoIter<K, V> {
-    pub fn new(vec: Vec<Vec<(K, V)>>) -> Self {
-        let mut outer = vec.into_iter();
-        let inner = { outer.next() }
-            .map(|v| v.into_iter())
-            .unwrap_or_else(|| (Vec::new()).into_iter());
-
-        IntoIter {
-            outer: outer,
-            inner: inner,
-        }
-    }
-}
-
-impl<K, V> Iterator for IntoIter<K, V> {
-    type Item = (K, V);
-
-    #[inline]
-    fn next(&mut self) -> Option<(K, V)> {
-        loop {
-            if let Some(r) = self.inner.next() {
-                return Some((r.0, r.1));
-            }
-
-            self.inner = self.outer.next()?.into_iter();
-        }
-    }
-}
-
-// ***************** Extend *********************
-
-impl<V> Extend<(u64, V)> for IntMap<V> {
-    #[inline]
-    fn extend<T: IntoIterator<Item = (u64, V)>>(&mut self, iter: T) {
-        for elem in iter {
-            self.insert(elem.0, elem.1);
-        }
-    }
-}
-
-// ***************** FromIterator *********************
-
-impl<V> std::iter::FromIterator<(u64, V)> for IntMap<V> {
-    #[inline]
-    fn from_iter<T: IntoIterator<Item = (u64, V)>>(iter: T) -> Self {
-        let iterator = iter.into_iter();
-        let (lower_bound, _) = iterator.size_hint();
-
-        let mut map = IntMap::with_capacity(lower_bound);
-        for elem in iterator {
-            map.insert(elem.0, elem.1);
-        }
-        map
-    }
-}
 
 // ***************** Equality *********************
